refactor: move marshall-contexts to smart-wallet
