+++ packages/wallet/api/package.json
@@ -28,6 +28,7 @@
     "@agoric/internal": "^0.2.1",
     "@agoric/nat": "^4.1.0",
     "@agoric/notifier": "^0.5.1",
+    "@agoric/smart-wallet": "^0.4.2",
     "@agoric/store": "^0.8.3",
     "@agoric/vat-data": "^0.4.3",
     "@agoric/zoe": "^0.25.3",
